+++ R/get_nasa_firms.R
@@ -46,7 +46,9 @@ NULL
                             instrument = "VIIRS",
                             rundir = tempdir(),
                             verbose = TRUE) {
+
   target_years <- attributes(x)$years
+
   if (!instrument %in% c("MODIS", "VIIRS")) {
     stop(
       paste("The selected instrument", instrument, "is not available. Please choose one of: MODIS or VIIRS")
@@ -56,12 +58,12 @@ NULL
   if (instrument == "VIIRS") {
     available_years <- c(2012:2021)
     target_years <- .check_available_years(
-      target_years, available_years, "active_fire"
+      target_years, available_years, "nasa_firms"
     )
   } else {
     available_years <- c(2000:2021)
     target_years <- .check_available_years(
-      target_years, available_years, "active_fire"
+      target_years, available_years, "nasa_firms"
     )
   }
 
@@ -70,13 +72,11 @@ NULL
     rundir,
     basename(paste0(instrument, "_", target_years, ".zip"))
   )
+
   if (attr(x, "testing")) {
     return(basename(filenames))
   }
 
-  if (any(file.exists(filenames))) {
-    message("Skipping existing files in output directory.")
-  }
   # start download in a temporal directory within tmpdir
   aria_bin <- attributes(x)$aria_bin
   .download_or_skip(urls, filenames, verbose, aria_bin = aria_bin)
@@ -86,49 +86,27 @@ NULL
   # remove unnecessary files
   d_files <- list.files(rundir, full.names = T)
   unlink(grep("_20*", d_files, value = T, invert = T),
-    recursive = T, force = T
+         recursive = T, force = T
   )
   # return paths to the gpkg
-  list.files(rundir, full.names = T, pattern = ".gpkg")
+  list.files(rundir, full.names = T, pattern = ".gpkg$")
 }
 
 
 #' A helper function to construct correct FIRMS layer urls
 #'
 #' @param target_year A numeric indicating the target year
-#' @param country A character vector specifying
-#'   the country name.
 #' @param instrument A character vector specifying the
 #'   data collection instrument.
 #'
-#' @return A character vector
+#' @return A character vector with the URL to the zip file of the respective year.
 #' @keywords internal
 #' @noRd
 .get_firms_url <- function(target_year, instrument) {
   if (instrument == "VIIRS") {
-    available_years <- c(2012:2021)
-    if (target_year %in% available_years) {
-      paste0("https://firms.modaps.eosdis.nasa.gov/data/country/zips/viirs-snpp_", target_year, "_all_countries.zip")
-    } else {
-      warning(
-        sprintf(
-          "FIRMS data from VIIRS S-NPP not available for target year %s", target_year
-        )
-      )
-      NULL
-    }
+    paste0("https://firms.modaps.eosdis.nasa.gov/data/country/zips/viirs-snpp_", target_year, "_all_countries.zip")
   } else {
-    available_years <- c(2000:2021)
-    if (target_year %in% available_years) {
-      paste0("https://firms.modaps.eosdis.nasa.gov/data/country/zips/modis_", target_year, "_all_countries.zip")
-    } else {
-      warning(
-        sprintf(
-          "FIRMS data from MODIS not available for target year %s", target_year
-        )
-      )
-      NULL
-    }
+    paste0("https://firms.modaps.eosdis.nasa.gov/data/country/zips/modis_", target_year, "_all_countries.zip")
   }
 }
 
@@ -139,8 +117,7 @@ NULL
 #' @param rundir The directory to where the files are unzipped
 #' @param instrument A character vector specifying the
 #'   data collection instrument.
-#' @importFrom base tolower
-#' @return Nothing.
+#' @return Nothing, its called for the side effect of unzipping.
 #' @keywords internal
 #' @noRd
 .unzip_firms <- function(zip, rundir, instrument) {
@@ -148,8 +125,9 @@ NULL
   year <- gsub(".*?([0-9]+).*", "\\1", bn)
 
   gpkg <- file.path(rundir, paste0("active_fire_", tolower(instrument), "_", year, ".gpkg"))
+
   if (file.exists(gpkg)) {
-    return(gpkg)
+    return()
   }
 
   utils::unzip(
@@ -162,6 +140,7 @@ NULL
 
   # bind all the CSVs and convert to gpkg
   .convert_csv_to_gpkg(
+    gpkg = gpkg,
     rundir = rundir,
     instrument = instrument,
     year = year
@@ -174,43 +153,28 @@ NULL
 #' @param rundir A directory where intermediate files are written to.
 #' @param instrument A character vector specifying the
 #'   data collection instrument.
+#' @param year An integer indicating the year of observation
+#' @param gpkg A character vector indicating the filename for the geopackage
 #'
 #' @importFrom utils read.csv
-#' @importFrom plyr rbind.fill
-#' @return Nothing.
+#' @importFrom purrr walk
+#' @return Nothing, its called for the side effect to create a gpkg.
 #' @keywords internal
 #' @noRd
 #'
-.convert_csv_to_gpkg <- function(rundir, instrument, year) {
+.convert_csv_to_gpkg <- function(gpkg, rundir, instrument, year) {
+
   if (instrument == "VIIRS") {
-
-    # bind all the CSVs
-    data_all <- suppressWarnings(list.files(paste0(rundir, "/viirs-snpp/", year, "/"),
-      pattern = "*.csv", full.names = TRUE
-    ) %>%
-      lapply(read.csv) %>%
-      plyr::rbind.fill())
-
+    loc <- file.path(rundir, "viirs-snpp", year)
   } else {
-
-    # bind all the CSVs
-    data_all <- suppressWarnings(list.files(paste0(rundir, "/modis/", year, "/"),
-      pattern = "*.csv", full.names = TRUE
-    ) %>%
-      lapply(read.csv) %>%
-      plyr::rbind.fill())
+    loc <- file.path(rundir, "modis", year)
   }
 
-  # convert to sf object
-  fire <- st_as_sf(
-    x = data_all,
-    coords = c("longitude", "latitude"),
-    crs = 4326
-  )
+  csv_files <- list.files(loc, pattern = "*.csv", full.names = TRUE)
 
-  # write fire as geopackage
-  write_sf(fire, file.path(
-    rundir,
-    paste0("active_fire_", tolower(instrument), "_", year, ".gpkg")
-  ))
+  walk(csv_files, function(file){
+    data <- read.csv(file)
+    data <- st_as_sf(data, coords = c("longitude", "latitude"), crs = 4326)
+    write_sf(data, dsn = gpkg, append = TRUE)
+  })
 }